update doc comment reply

+++ backend/Blog/views.py
@@ -1,3 +1,5 @@
+from ast import Param
+from inspect import Parameter
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
@@ -160,6 +162,24 @@ def getAllPost(request):
 
 
 # -----------------------------------------------------------------------
+@swagger_auto_schema(
+    method='post',
+    request_body=openapi.Schema(
+        type=openapi.TYPE_OBJECT,
+        properties={
+            'content': openapi.Schema(type=openapi.TYPE_STRING)
+        }
+    ),
+    manual_parameters=[
+        openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                          type=openapi.TYPE_STRING),
+        openapi.Parameter('slug', in_=openapi.IN_PATH,
+                          description='slug Post', type=openapi.TYPE_STRING)
+    ],
+    responses={
+        200: 'Success'
+    }
+)
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def addComment(request, slug):
@@ -175,40 +195,83 @@ def addComment(request, slug):
 # -----------------------------------------------------------------------
 
 
+@swagger_auto_schema(
+    method='delete',
+    manual_parameters=[
+        openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                          type=openapi.TYPE_STRING),
+        openapi.Parameter('id', in_=openapi.IN_PATH, type=openapi.TYPE_STRING)
+
+    ],
+    responses={
+        200: 'Success',
+        401: 'UNAUTHORIZED'
+    }
+
+)
 @api_view(['DELETE'])
 @permission_classes([IsAuthenticated])
 def deleteComment(request, id):
     user = request.user
     comment = models.Comment.objects.get(id=id)
     if user.id != comment.user.id:
-        return Response(data='Fail', status=status.HTTP_401_UNAUTHORIZED)
+        return Response(data='UNAUTHORIZED', status=status.HTTP_401_UNAUTHORIZED)
     comment.delete()
     return Response(data='Success', status=status.HTTP_200_OK)
 
 # -----------------------------------------------------------------------
 
 
+@swagger_auto_schema(
+    method='put',
+    request_body=openapi.Schema(
+        type=openapi.TYPE_OBJECT,
+        properties={
+            'content': openapi.Schema(type=openapi.TYPE_STRING)
+        }
+    ),
+    manual_parameters=[
+        openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                          type=openapi.TYPE_STRING),
+        openapi.Parameter('id', in_=openapi.IN_PATH, type=openapi.TYPE_STRING)
+    ],
+    responses={
+        200: 'Success',
+        401: 'UNAUTHORIZED'
+    }
+
+)
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
 def editComment(request, id):
     user = request.user
-    comment = models.Comment.objects.get(id=id)
-    if user.id != comment.user.id:
+    comment = models.Comment.objects.filter(id=id)
+    if user.id != comment[0].user.id:
         return Response(data='Fail', status=status.HTTP_401_UNAUTHORIZED)
-
-    rs = serializers.CommentSerializer(instance=comment, data={
-        'user': user.id,
-        'post': comment.post.id,
-        'content': request.data['content']
-    })
-    if rs.is_valid(raise_exception=True):
-        rs.save()
-        return Response(data=rs.data, status=status.HTTP_200_OK)
-    return Response(data='Fail', status=status.HTTP_400_BAD_REQUEST)
+    comment.update(content=request.data['content'])
+    return Response(data='Success', status=status.HTTP_200_OK)
 
 # -----------------------------------------------------------------------
 
 
+@swagger_auto_schema(
+    method='post',
+    request_body=openapi.Schema(
+        type=openapi.TYPE_OBJECT,
+        properties={
+            'content': openapi.Schema(type=openapi.TYPE_STRING)
+        }
+    ),
+    manual_parameters=[
+        openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                          type=openapi.TYPE_STRING),
+        openapi.Parameter('id', description='id comment',
+                          in_=openapi.IN_PATH, type=openapi.TYPE_STRING)
+    ],
+    responses={
+        200: 'Success'
+    }
+)
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def addReply(request, id):
@@ -224,24 +287,57 @@ def addReply(request, id):
 
 
 # -----------------------------------------------------------------------
+@swagger_auto_schema(
+    method='put',
+    request_body=openapi.Schema(
+        type=openapi.TYPE_OBJECT,
+        properties={
+            'content': openapi.Schema(type=openapi.TYPE_STRING)
+        }
+    ),
+    manual_parameters=[
+        openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                          type=openapi.TYPE_STRING),
+        openapi.Parameter('id', description='id reply',
+                          in_=openapi.IN_PATH, type=openapi.TYPE_STRING)
+    ],
+    responses={
+        200: 'Success',
+        401: 'UNAUTHORIZED'
+    }
+)
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
 def editReply(request, id):
     user = request.user
     reply = models.Reply.objects.filter(id=id)
-    # if user.id != reply.user.id:
-    #     return Response(data='Fail', status=status.HTTP_401_UNAUTHORIZED)
+    if user.id != reply[0].user.id:
+        return Response(data='UNAUTHORIZED', status=status.HTTP_401_UNAUTHORIZED)
     reply.update(content=request.data['content'])
     return Response(data='Success', status=status.HTTP_200_OK)
-    # return Response(data='Fail', status=status.HTTP_400_BAD_REQUEST)
 
 
 # -----------------------------------------------------------------------
 
+@swagger_auto_schema(
+    method='delete',
 
+    manual_parameters=[
+        openapi.Parameter('Authorization', in_=openapi.IN_HEADER,
+                          type=openapi.TYPE_STRING),
+        openapi.Parameter('id', description='id reply',
+                          in_=openapi.IN_PATH, type=openapi.TYPE_STRING)
+    ],
+    responses={
+        200: 'Success',
+        401: 'UNAUTHORIZED'
+    }
+)
 @api_view(['DELETE'])
 @permission_classes([IsAuthenticated])
 def deleteReply(request, id):
     reply = models.Reply.objects.get(id=id)
+    if request.user.id != reply.user.id:
+        return Response(data='UNAUTHORIZED', status=status.HTTP_401_UNAUTHORIZED)
     reply.delete()
     return Response(data='Success', status=status.HTTP_200_OK)
